### 2024-09-19 #89 @rochanaro
[Re] Network Deconvolution
**Original article:** C. Ye, M. Evanusa, H. He, A. Mitrokhin, T. Goldstein, J. A. Yorke, C. Fermüller, and Y. Aloimonos. “Network
Deconvolution.” In: ICLR (2020).

**Scientific domain:**  Machine Learning
