### 2025-01-21 #89 @rougier
Very sorry for such a long delay, hopefully things will get better for 2025.
I was asking the question because the format of the PDF is very similar to the ICLR challenge. Note that this is not a problem at all, the idea was to re-use review if they were available.

I'll edit your review and assign reviewers soon hopefully.

In the meantime, can you have a look at [other submissions](https://github.com/ReScience/submissions/issues) and propose yourself to review?
