### 2025-01-23 #89 @rochanaro
> Very sorry for such a long delay, hopefully things will get better for 2025. I was asking the question because the format of the PDF is very similar to the ICLR challenge. Note that this is not a problem at all, the idea was to re-use review if they were available.
>
> I'll edit your review and assign reviewers soon hopefully.
>
> In the meantime, can you have a look at [other submissions](https://github.com/ReScience/submissions/issues) and propose yourself to review?

Thank you for the update! We understand how busy things can get.
Yes, for organizing the content, we tried to follow the structure of MLRC 2022 template only to ensure clarity.
I’ll certainly take a look at other submissions and will propose myself as a reviewer where I can contribute.
