### 2025-02-17 #89 @alchemi5t
@rougier I would like to review this one.
