### 2025-03-12 #89 @jsta
@rougier Yes, I have some general comments prepared. I will post them once I have finished a couple tests to verify general behavior. I do not plan to run all the tests due to computational constraints.
