### 2025-03-17 #89 @alchemi5t
Dear Authors,

Here's my review of your work.

# Review of "[Re] Network Deconvolution"

This reproduction study provides a thorough evaluation of the network deconvolution technique introduced by Ye et al. (2020). After examining both the original paper and this reproduction, I find that the authors have validated the primary claim about network deconvolution for most parts. While there are many cases where network deconvolution improves model performance compared to batch normalization, the reproduction results show few exceptions that contradict the original paper's universal claim. For example, in the ResNet-18 architecture with CIFAR-100 at 100 epochs, batch normalization (97.42%) actually outperformed network deconvolution (94.31%), which contradicts both the original paper's results and its central claim.

## Strengths

- **Comprehensive testing**: The authors tested 10 modern neural network architectures on CIFAR-10/100 and ImageNet datasets, over 3 runs
- **Training time analysis**: The authors went beyond the original paper by analyzing computational overhead, showing that network deconvolution requires more training time

## Key Observations

The few exceptions where BN seems to have outperformed has not been noted and there is no analysis around it. Furthermore, this statement is not true given the results:
"The results show that the model performance with network deconvolution is always better than the model performance with batch normalization."

Another notable finding is that the reproduced accuracy values were often significantly higher than those reported in the original paper. For example:

- For VGG-16 with CIFAR-100 at 100 epochs, the original paper reported 75.32% accuracy with network deconvolution, while the reproduction achieved 99.30%
- Similar large improvements were observed across most architectures and datasets

The authors attribute this systematic improvement to:
1. Advancements in numerical stability of libraries (NumPy 1.16.1 → 1.23.5, PyTorch 1.0 → 1.13)
2. Improved optimization algorithms and parallelism in Tensorflow

While these explanations are plausible, the magnitude of improvement (sometimes exceeding 20%) suggests there might be additional factors at play that weren't fully investigated. Given that this improvement has increased most accuracy (both ND and BN) to 99.xx where the comparison comes down to 10e-2 (vs the 1-2 pts in the original paper), I expected deeper analysis and stronger evidence for these claims.

## Code

As far as I understand, running training runs seems to be the only way to test the codebase. Due to lack of hardware I am unable to do so, but to anyone who can and would like to quickly verify the claims, releasing the trained weights at 100 epochs and a minimal script to infer the results would be greatly helpful.

## Opinion

This reproduction study provides valuable insights but reveals important discrepancies with the original paper's claims. While network deconvolution often outperforms batch normalization, the reproduction found notable exceptions.

The reproduction yielded substantially higher accuracy values for both techniques compared to those reported in the original paper. These significant discrepancies make it difficult to draw direct comparisons with the original results, and the proposed explanations for these differences (library improvements, optimization algorithms) remain speculative without rigorous empirical validation.

Rather than enhancing confidence in network deconvolution as a universal improvement over batch normalization, this reproduction suggests a more nuanced view: network deconvolution appears to be a viable alternative that performs better in many but not all scenarios. The authors' detailed reporting of computational costs and performance characteristics across architectures provides essential practical context for researchers considering which normalization technique to employ.
