### 2025-03-19 #89 @jsta
Ok, I have completed my review. Please see the comments below. In addition to reading the paper and git repository, I verified that I could run 2 arbitrary architectures (vgg16 and pnasnetA) for 100 epochs for both the BN and ND cases, and got similar relative accuracy results.

### major comments
* Section 1, last paragraph, says "our study attempts to reproduce the results reported in the original paper with the most recent versions of software libraries.". This is not accurate, for example, Tensorflow 2.12 is years old. Maybe change to "the most recent version of software libraries as of [some date]"?

### minor comments
* Section 4.1, first sentence, says "compared it against the state‐of‐the‐art against". This seems like a typo or grammar mistake.
* Is there a typo in `imagenet_single_experiment_densenet121.sh`, should `-a densenet121` -> `-a densenet121d`?

### highly optional and/or minor curiosity questions
* Section 5.1, first paragraph, says "[...] we averaged the results of three attempts and compared them with the original study’s reported values.". Was there variation among the attempts such that running each model/architecture combination that many times was worthwhile?
* Figure 3, why were the reproduced values so much more accurate than the original for CIFAR100 but not CIFAR10 or Imagenet?
* There is some information in the repository README that duplicates the paper. I wonder if you could remove information about the steps you did to create the repo and focus the README on how users will interact with it. Get them up and running quickly without wading through extraneous details. Conversely, the papers says that torchvision datasets are downloaded on-the-fly but this is not in the README.
* It would be nice to add a time argument to the sbatch files so users have a rough idea of how long they're expected to run.
* It's strange that Table 8 lists out the model names while Table 4 keys them out by number code.
