### 2025-03-24 #89 @rougier
@alchemi5t @jsta Many  thanks for your detailed reviews.
@rochanaro Can you address the commend and specifically address @alchemi5t concerns about performances? It seems the change in numeric stability might not be a plausible reason for enhanced performances. Any idea on that ? Would it be possible to rerun your code (without too much hassle) with the exact same stack as the original paper?
