### 2025-03-24 #89 @rochanaro
> [@alchemi5t](https://github.com/alchemi5t) [@jsta](https://github.com/jsta) Many thanks for your detailed reviews. [@rochanaro](https://github.com/rochanaro) Can you address the commend and specifically address [@alchemi5t](https://github.com/alchemi5t) concerns about performances? It seems the change in numeric stability might not be a plausible reason for enhanced performances. Any idea on that ? Would it be possible to rerun your code (without too much hassle) with the exact same stack as the original paper?

Thank you @alchemi5t and @jsta for the reviews. @rougier It would be challenging to redo the experiment using the exact same dependencies as the original paper. During our study, we attempted to contact the original authors to obtain the precise library versions, as the original study repository does not mention them, but we did not receive a response. Therefore, we adopted the following approach: we used the latest available versions of each library as of 2020 and only opted for more recent versions when compatibility issues arose.

We will soon get back with answers, revisions, and updates addressing the above concerns in our upcoming response.
