### 2025-04-01 9872dcb
test scripts added for direct evaluation (without training) - for cifar10, cifar100 and imagenet datasets
