+++ test_script_for_table_1.py
@@ -1,0 +1,275 @@
+from __future__ import print_function
+import torch
+import time
+import os
+import shutil
+from torchvision import models, datasets, transforms
+# from torch.utils.data import DataLoader
+import torch.nn.functional as F
+from functools import partial
+import torch.nn as nn
+from models import *
+from net_util import *
+from arg_parser import *
+
+
+# Helper class for tracking metrics
+class AverageMeter:
+    def __init__(self):
+        self.reset()
+
+    def reset(self):
+        self.val = 0
+        self.avg = 0
+        self.sum = 0
+        self.count = 0
+
+    def update(self, val, n=1):
+        self.val = val
+        self.sum += val * n
+        self.count += n
+        self.avg = self.sum / self.count
+
+# Accuracy computation function
+def accuracy(output, target, topk=(1,)):
+    """Computes the accuracy over the k top predictions for the specified values of k"""
+    with torch.no_grad():
+        maxk = max(topk)
+        batch_size = target.size(0)
+
+        _, pred = output.topk(maxk, 1, True, True)
+        pred = pred.t()
+        correct = pred.eq(target.view(1, -1).expand_as(pred))
+
+        res = []
+        for k in topk:
+            correct_k = correct[:k].reshape(-1).float().sum(0, keepdim=True)
+            res.append(correct_k.mul_(100.0 / batch_size))
+        return res
+
+# Load dataset
+def load_dataset(dataset_name, args, batch_size=128):
+
+    if args.dataset=='cifar10':
+        args.in_planes = 3
+        args.input_size=32
+
+        transform_test = transforms.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize((0.4914, 0.4822, 0.4465), (0.2023, 0.1994, 0.2010)),
+        ])
+        args.classes = ('plane', 'car', 'bird', 'cat',
+                        'deer', 'dog', 'frog', 'horse', 'ship', 'truck')
+
+        print("| Preparing CIFAR-10 dataset...")
+        testset = torchvision.datasets.CIFAR10(root='./data', train=False, download=True, transform=transform_test)
+        args.num_outputs = 10
+        classes = ('plane', 'car', 'bird', 'cat', 'deer', 'dog', 'frog', 'horse', 'ship', 'truck')
+
+    elif (args.dataset == 'cifar100'):
+        args.in_planes = 3
+        args.input_size = 32
+
+        transform_test = transforms.Compose([
+            transforms.ToTensor(),
+            transforms.Normalize((0.5071, 0.4865, 0.4409), (0.2009, 0.1984, 0.2023)),
+        ])
+
+        print("| Preparing CIFAR-100 dataset...")
+        testset = torchvision.datasets.CIFAR100(root='./data', train=False, download=True, transform=transform_test)
+        args.num_outputs = 100
+        
+    test_loader = torch.utils.data.DataLoader(testset, batch_size=args.batch_size, shuffle=True, num_workers=args.num_workers)
+    
+    return test_loader
+
+
+# Load model checkpoint
+def load_model(net, model_path):
+    checkpoint = torch.load(model_path, map_location=torch.device('cpu'))
+   
+    # model = getattr(models, architecture)(pretrained=False)
+    net.load_state_dict(checkpoint['state_dict'], strict=False)  # Allow mismatches if minor
+    net.eval()
+
+    return net
+
+
+# Evaluation function (from your `eval_net`)
+def eval_net(net, data_loader, use_gpu=False):
+    top1 = AverageMeter()
+    top5 = AverageMeter()
+    losses = AverageMeter()
+
+    net.eval()
+    total_time = 0
+    end_time = time.time()
[…]
