### 2025-04-09 #89 @rochanaro
# Full response to ReScience C reviews (submission #89)

Dear editor (@rougier) and the reviewers (@alchemi5t, @jsta) ,

We appreciate the reviewers’ careful evaluation of our reproduction study. Below is our detailed response addressing each of the reviewers’ points:

### Performance Discrepancies and Exceptional Cases

> While there are many cases where network deconvolution improves model performance compared to batch normalization, the reproduction results show few exceptions that contradict the original paper's universal claim. For example, in the ResNet-18 architecture with CIFAR-100 at 100 epochs, batch normalization (97.42%) actually outperformed network deconvolution (94.31%), which contradicts both the original paper's results and its central claim.

We appreciate the observation that batch normalization (BN) outperformed network deconvolution in certain cases (e.g., ResNet-18 on CIFAR-100 at 100 epochs), and we acknowledge these exceptions. Our study aimed to carefully reproduce the experiments rather than optimize hyperparameters or modify architectural choices. In the literature, such variations are not uncommon when replicating older experiments (see, e.g., [Examining the Effect of Implementation Factors on Deep Learning Reproducibility](https://www.computer.org/csdl/proceedings-article/e-science/2022/612400a397/1J6hpWv0a5O), [A Study on Reproducibility and Replicability of Table Structure Recognition Methods](https://link.springer.com/chapter/10.1007/978-3-031-41679-8_1), [Towards training reproducible deep learning models](https://dl.acm.org/doi/10.1145/3510003.3510163)). We do not claim that network deconvolution universally outperforms BN; rather, our findings suggest that while ND often provides benefits, its effectiveness may be context dependent (explained in manuscript section 6 paragraph 4).

### Systematic Accuracy Improvements

> Another notable finding is that the reproduced accuracy values were often significantly higher than those reported in the original paper. For example:
•	For VGG-16 with CIFAR-100 at 100 epochs, the original paper reported 75.32% accuracy with network deconvolution, while the reproduction achieved 99.30%
•	Similar large improvements were observed across most architectures and datasets

We acknowledge the reviewer’s observation regarding systematically higher accuracy values (e.g., VGG-16 with CIFAR-100) and agree on the importance of understanding these differences. We believe that several factors contribute to these improvements. Notably, advancements in numerical stability, as documented in recent studies (e.g., [Recent advances and applications of deep learning methods in materials science](https://www.nature.com/articles/s41524-022-00734-6), [A Comprehensive Review of Deep Learning: Architectures, Recent Advances, and Applications](https://www.mdpi.com/2078-2489/15/12/755) ), and enhancements in optimization algorithms may play a significant role.

Additionally, prior research supports the idea that updated frameworks can improve model performance. [Mienye et al. (2024)](https://www.mdpi.com/2078-2489/15/12/755), and [Qin et al. 2024](https://www.eurekalert.org/news-releases/1063122) and [Vaishnav et al. (2022)](https://www.nature.com/articles/s41586-022-04506-6)  demonstrated that deep learning models consistently achieve higher accuracy when trained with updated frameworks due to improvements in optimization techniques and regularization strategies. Similarly, [Coakley et al. (2023)](https://doi.ieeecomputersociety.org/10.1109/eScience55777.2022.00056) found that updated hardware and software environments alone can introduce accuracy differences of over 6%, highlighting the impact of implementation-level changes on reproducibility.

While the observed accuracy gains (e.g., 75% → 99% for VGG-16 on CIFAR-100) are substantial, we emphasize that both network deconvolution (ND) and batch normalization (BN) baselines improved comparably, suggesting that these gains are driven by systemic factors rather than being specific to any particular method. Although we have provided plausible explanations based on updates in software libraries, we acknowledge that these factors may not fully account for the magnitude of the observed differences. Further investigation into these aspects is warranted, but we reiterate that the primary objective was to assess reproducibility rather than to deconstruct every underlying cause of performance variation.

We provided detailed explanations in Section 6 Paragraphs 2 and 3 in the revised manuscript.

### Sharing trained weights for verification

> As far as I understand, running training runs seems to be the only way to test the codebase. Due to lack of hardware I am unable to do so, but to anyone who can and would like to quickly verify the claims, releasing the trained weights at 100 epochs and a minimal script to infer the results would be greatly helpful.

We acknowledge the reviewer's (@alchemi5t) request and have taken steps to facilitate easy verification of our results. The **trained model weights at 100 epochs** have been made publicly available at https://osf.io/hp3ab/files/osfstorage , corresponding to the results presented in Tables 1 and 2. Additionally, we provide two minimal inference scripts (`test_script_for_table_1.py` and `test_script_for_table_2.py`) that allow users to reproduce our reported accuracy without requiring extensive computational resources. This ensures that our findings can be easily validated by downloading the weights and running a single command.

The steps have been explained in the newly added section, `"To Validate the Results Using the Trained Weights (Direct Inference Without Training)"` in the GitHub README.md file and in Section 5.4 of the revised manuscript.

For example, the following command can be used to verify results for the VGG-16 architecture on CIFAR-10 with batch normalization:
```
python test_script_table_1.py --arch vgg16 --dataset cifar10 --deconv False --model_path "checkpoints/cifar10_vgg16_BN.pth.tar"
```

### Clarification on Software Versions (Reviewer 2’s Comment @jsta)

> Section 1, last paragraph, says "our study attempts to reproduce the results reported in the original paper with the most recent versions of software libraries.". This is not accurate, for example, Tensorflow 2.12 is years old. Maybe change to "the most recent version of software libraries as of [some date]"?

In response to Reviewer 2’s remark regarding our statement in Section 1, we have revised the language. The original phrase “with the most recent versions of software libraries” has been updated to “with the most recent versions of software libraries as of 2020” to prevent any misunderstanding (Section 1 Page 3). During our study, we attempted to contact the original authors to obtain the precise library versions used in their experiments, as the repository did not include this information. Unfortunately, we did not receive a response. Therefore, we adopted the following approach: we used the latest available versions of each library as of 2020 and opted for more recent versions only when compatibility issues arose. The details are explained in Section 4.5 of the revised manuscript.

# Summary and a list of changes made based on the reviewers’ and editor’s comments

***All the updates in the manuscript are highlighted in color blue.**

#### Reviewer 1 (@alchemi5t):

- Made trained weights public and introduced two scripts for direct model inference without training (revised manuscript Section 5.4)
- Explained our main objective and provided explanation on accuracy discrepancies (revised manuscript Section 6)

#### Reviewer 2 (@jsta):

- major comments:
  - revised the sentence in the manuscript Section 1 on Page 3.
- minor comments:
  - grammar mistake corrected (Section 4.1)
  - **-a densenet121** in GitHub file `imagenet_single_experiment_densenet121.sh` isn’t a typo
-  highly optional and/or minor curiosity questions:
   - The results were similar for all attempts, so the variation was minimal.
   - Since these are three different datasets, dataset-inherent factors may play a role. However, we suspect that the higher accuracy in the reproduced results for CIFAR-100 compared to CIFAR-10 and ImageNet is primarily due to improvements in regularization techniques and optimization methods in newer versions of PyTorch. (revised manuscript Section 6)
   - Mentioned about downloading CIFAR-10 and CIFAR-100 datasets via torchvision datasets in the README.md section `“To reproduce results of our reproducibility study”`  list Item 2
   - Added `--time` parameter for sbatch commands (updated in README.md section “Steps we have followed to reproduce the original study” list Items 5 and 6)
   - The intention was to include model names in the columns, but due to space constraints in Tables 4 and 5, we had to use number codes instead

#### Editor (@rougier):
- Explained why it is not possible to use the exact same software stack of the original study with our reproducibility study (explained in the revised manuscript Section 4.5 and in this response)
- Provided an explanation to Reviewer 1’s accuracy discrepancies (revised manuscript Section 6 newly added content in blue and in this GitHub response)

-----

Revised Manuscript URL: https://github.com/lamps-lab/rep-network-deconvolution/blob/master/article.pdf
Metadata URL: https://github.com/lamps-lab/rep-network-deconvolution/blob/master/metadata.yaml
Code URL: https://github.com/lamps-lab/rep-network-deconvolution
