### 2025-04-17 #89 @rougier
Great, then we can accept your replication! Congratulations.

For the actual publicaton, I would need a link, to a github repo with the sources of your articles. And you'll need to save you code repo on software heritage such as to get a SWID.
