### 2025-04-17 b06aac9
manuscript source files added
