+++ documents/manuscript/article.tex
@@ -1,0 +1,18 @@
+\documentclass{rescience}
+
+\begin{document}
+
+\input{header.tex}
+\vspace{-0.95cm}
+% Add the Reproducibility Summary in summary.tex
+\input{summary.tex}
+\newpage
+% Add your paper contents in content.tex
+\input{content.tex}
+
+\hypersetup{linkcolor=black,urlcolor=darkgray}
+\renewcommand\emph[1]{{\bfseries #1}}
+\setlength\bibitemsep{0pt}
+\printbibliography
+
+\end{document}

[…]
