### 2025-04-17 #89 @rochanaro
Thank you, @rougier. The requested items are listed below. Please let us know if anything additional is needed.

GitHub link for the article source files: https://github.com/lamps-lab/rep-network-deconvolution/tree/master/documents/manuscript
SWID:  swh:1:dir:680efae9c7c628f93ee03731e73c34cf00b0e245

swh:1:dir:680efae9c7c628f93ee03731e73c34cf00b0e245;
origin=https://github.com/lamps-lab/rep-network-deconvolution;
visit=swh:1:snp:6946519d5ef5dd5312c1da6543218e512c2b0633;
anchor=swh:1:rev:b06aac9c1ea4bf38d2dd86eaf63b7dd24f10f561
