### 2025-04-24 #89 @rougier
Sorry I'm late in handling your article. Can you fill in the metadata.yaml with as much information as you can ? By the way, you may need to take the new template that is available here: https://github.com/ReScience/template (with a swh entry).

For editor/reviewers, you'll find orcid numbers  on https://rescience.github.io/board/
