### 2025-04-25 28ada70
article template updated with SWH entry
