+++ documents/manuscript/article.tex
@@ -4,6 +4,7 @@
 
 \input{header.tex}
 \vspace{-0.95cm}
+\vspace{-10pt}
 % Add the Reproducibility Summary in summary.tex
 \input{summary.tex}
 \newpage
