### 2025-04-25 #89 @rochanaro
Dear Editor (@rougier),

We have updated the metadata.yaml file and the template with SWH entry. Please let us know if there is anything additional we can do.

Manuscript source files: https://github.com/lamps-lab/rep-network-deconvolution/tree/master/documents/manuscript
metadata.yaml : https://github.com/lamps-lab/rep-network-deconvolution/blob/master/documents/manuscript/metadata.yaml
Article PDF: https://github.com/lamps-lab/rep-network-deconvolution/blob/master/article.pdf

Thank you!
