### 2025-04-28 #89 @rougier
Thanks! Sandbox version is online at https://sandbox.zenodo.org/record/203035. Can you check everything is ok (including links pointing to the right place) and tell me such that I can publish the final version?
