### 2025-05-02 #89 @rougier
Great, thanks. Your paper is online at https://zenodo.org/record/15172014/ and will soon appear on ReScience website.
Congratulations!
