### 2025-05-02 #89 @rougier
Sorry, I meant https://zenodo.org/records/15321683
